+++ lib/segment/src/index/hnsw_index/tests/test_graph_connectivity.rs
@@ -62,7 +62,7 @@ fn test_graph_connectivity() {
     };
 
     let permit_cpu_count = num_rayon_threads(hnsw_config.max_indexing_threads);
-    let permit = Arc::new(ResourcePermit::dummy(permit_cpu_count as u32));
+    let permit = Arc::new(ResourcePermit::dummy(permit_cpu_count));
 
     let hnsw_index = HNSWIndex::build(
         HnswIndexOpenArgs {
