+++ lib/segment/tests/integration/segment_on_disk_snapshot.rs
@@ -139,7 +139,7 @@ fn test_on_disk_segment_snapshot(#[case] format: SnapshotFormat) {
     let mut rng = rand::rng();
     let segment = segment_builder
         .build(
-            ResourcePermit::dummy(num_rayon_threads(0) as u32),
+            ResourcePermit::dummy(num_rayon_threads(0)),
             &false.into(),
             &mut rng,
             &HardwareCounterCell::new(),
